+++ src/test/java/com/luq/store/controllers/CSV/SupplyCSVControllerTest.java
@@ -17,6 +17,7 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.Locale;
@@ -63,11 +64,11 @@ public class SupplyCSVControllerTest {
 
         fakeProduct1 = pRepository.save(new Product(
             "XOneCont", "Xbox One Controller", "Controller for Xbox One Console",
-            fakeSupplier1, 200.00F
+            fakeSupplier1, BigDecimal.valueOf(200.00)
         ));
         fakeProduct2 = pRepository.save(new Product(
             "PS5Cont", "Playstation 5 Controller", "Controller for Playstation 5 Console",
-            fakeSupplier2, 250.00F
+            fakeSupplier2, BigDecimal.valueOf(250.00)
         ));
 
         fakeSupply1 = new Supply(20, fakeProduct1);
