+++ src/test/java/com/luq/store/repositories/OrderRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.luq.store.repositories;
 
 import com.luq.store.domain.*;
+import com.luq.store.domain.Customer;
 import com.luq.store.valueobjects.Cnpj;
 import com.luq.store.valueobjects.Mail;
 import com.luq.store.valueobjects.Phone;
@@ -12,6 +13,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.Sort;
 import org.springframework.test.annotation.DirtiesContext;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -54,11 +56,11 @@ public class OrderRepositoryTest {
 
         fakeProduct1 = pRepository.save(new Product(
             "XOneCont", "Xbox One Controller", "Controller for Xbox One Console",
-            fakeSupplier1, 200.00F
+            fakeSupplier1, BigDecimal.valueOf(200.00)
         ));
         fakeProduct2 = pRepository.save(new Product(
             "PS5Cont", "Playstation 5 Controller", "Controller for Playstation 5 Console",
-            fakeSupplier2, 250.00F
+            fakeSupplier2, BigDecimal.valueOf(250.00)
         ));
 
         fakeSeller1 = sellerRepository.save(new Seller(
@@ -71,10 +73,10 @@ public class OrderRepositoryTest {
         ));
 
         fakeOrder1 = oRepository.save(new Order(
-            400.00F, 5, LocalDate.now(), fakeProduct1, fakeSeller1, fakeCustomer1
+            BigDecimal.valueOf(400.00F), 5, LocalDate.now(), fakeProduct1, fakeSeller1, fakeCustomer1
         ));
         fakeOrder2 = oRepository.save(new Order(
-            800.00F, 8, LocalDate.now(), fakeProduct2, fakeSeller2, fakeCustomer2
+            BigDecimal.valueOf(800.00F), 8, LocalDate.now(), fakeProduct2, fakeSeller2, fakeCustomer2
         ));
     }
