+++ src/test/java/com/luq/store/services/SupplyServiceTest.java
@@ -15,6 +15,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -49,11 +50,11 @@ public class SupplyServiceTest {
 
         Product fakeProduct1 = new Product(
                 1, "Xbox One Controller", "XOneCont", "Controller for Xbox One Console",
-                200.00F, fakeSupplier1, user, now, user, now
+                BigDecimal.valueOf(200.00), fakeSupplier1, user, now, user, now
         );
         Product fakeProduct2 = new Product(
                 2, "Playstation 5 Controller", "PS5Cont", "Controller for Playstation 5 Console",
-                250.00F, fakeSupplier2, user, now, user, now
+                BigDecimal.valueOf(250.00), fakeSupplier2, user, now, user, now
         );
 
         fakeSupply1 = new Supply(1, 100, fakeProduct1, user, now, user, now);
